observability: allow Span with key for hotkey detection
Hotkey detection can now be performed on your backend
by corelating a specific key with a latency.
A hotkey is one that causes a pathologic system
response time e.g. if the key "foo" causes
the system to return in 10ms yet every other
key returns in about 570us.

With a sufficiently expressive filtering backend,
you can ask for the hotest key i.e. with the most
latency.

The last problem though is that we aren't tracking
their frequencies, but that perhaps isn't the point
of tracking hotkeys and has no relation.

The key to filter on is "key" and we can further
drill down by "command" since that's another tag
that's used.